Add support for Klingon character drop
Adds character set option that for the Klingon alphabet with argument '-l K'. The used characters are part of the Unicode standard given to Klingon characters, which is a subset of the "ConScripts". For it to work, the user's terminal font must implement the Klingon characters. Works with "FreeMono Regular" in `mate-terminal` on Ubuntu 17.10.